Commit for section 18: Unit of work patter and finishinh touches

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -16,7 +17,7 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messagesService: MessageService) { }
+  constructor(private messagesService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -33,8 +34,21 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(message: Message) {
-    this.messagesService.deleteMessage(message.id.toString()).subscribe(() => {
-      this.messages = this.messages.filter(m => m.id !== message.id);
-    });
+    this.confirmService.popUpConfirm('Confirm delete message', 'This cannot be undone').subscribe(
+      result => {
+        if(result) {
+          this.messagesService.deleteMessage(message.id.toString()).subscribe(() => {
+            this.messages = this.messages.filter(m => m.id !== message.id);
+          });
+        }
+      }
+    )
+  }
+
+  pageChanged(event: any) {
+    if (this.pageNumber !== event.page) {
+      this.pageNumber = event.page;
+      this.loadMessages();
+    }
   }
 }
